+++ app.js
@@ -75,6 +75,6 @@ app.use((req, res, next) => {
 
 app.listen(port, () => console.log(`Express app listening on port ${port}!`));
 
-require("./routes")(app, passport);
+require("./routes")(app);
 
 module.exports = app;
