+++ routes/apis.js
@@ -1,0 +1,9 @@
+// import router module
+const express = require('express')
+const router = express.Router()
+
+router.get('/', (req, res) => {
+  return res.json('this is apis')
+})
+
+module.exports = router
